Test refactoring
